test: rm unwantes tools and utils
