+++ pyproject.toml
@@ -84,7 +84,7 @@ addopts = ["-p no:warnings", "--import-mode=importlib"]
 xfail_strict = true
 log_cli_level = "warn"
 pythonpath = "src"
-testpaths = ["tests"]
+testpaths = ["tests", "tasks/catalyst/tests"]
 
 [tool.black]
 exclude = '''
