+++ tasks/catalyst/src/catalyst/tools.py
@@ -3,19 +3,14 @@ import os
 import subprocess
 import sys
 import tempfile
-import traceback
 from pathlib import Path
 
 from catalyst.tool_utils import (
     ensure_directory_exists,
-    find_surface_atoms_with_voronoi,
     generate_output_capture_code,
     load_structure,
     parse_execution_output,
     safe_convert_timeout,
-    set_fixed_atom_constraints,
-    standardize_bulk,
-    tile_atoms,
 )
 from dotenv import load_dotenv
 from loguru import logger
@@ -93,26 +88,6 @@ def get_bulk_polymorphs_data_func(composition: str) -> str:
         polymorph_data = sorted(polymorph_data, key=lambda x: x["energy_above_hull"])
 
         return json.dumps(polymorph_data, indent=2)
-
-
-def execute_python_code_given_code(code: str) -> str:
-    """
-    Executes a given Python code string.
-    This is a placeholder and should be replaced with your actual implementation.
-    """
-    try:
-        # Create a dictionary to hold local variables during execution
-        exec_globals = {}
-        exec_locals = {}
-        exec(code, exec_globals, exec_locals)
-        # Assuming the filtering code will produce a 'output' variable
-        return json.dumps(
-            {"success": True, "execution_result": {"output": exec_locals.get("output")}}
-        )
-    except Exception as e:
-        return json.dumps(
-            {"success": False, "error": str(e), "traceback": traceback.format_exc()}
-        )
 
 
 ####################
@@ -1849,7 +1824,7 @@ def execute_python_script(
     [/LIMITATIONS]
     """
     try:
-        if not Path.exists(script_path):
+        if not Path(script_path).exists():
             return json.dumps(
                 {"success": False, "error": f"Script file not found: {script_path}"}
             )
@@ -2301,856 +2276,6 @@ def get_mp_surface_properties(material_id: str) -> str:
             return json.dumps({"error": f"Error fetching surface properties: {e!s}"})
 
 
-########################
-# Extra  tools for ablations
-########################
-
-
-@tool
-def process_slab_ocdata_style(
-    slab_cif: str,
-    bulk_cif: str,
-    min_xy_size: float = 8.0,
-    apply_constraints: bool = True,
-) -> str:
-    """[BRIEF] Process a raw slab structure with ocdata-style surface atom tagging, tiling, and constraints for computational surface studies. [/BRIEF]
-
-    [DETAILED] This tool applies a comprehensive processing pipeline to raw slab structures, implementing the ocdata
-    methodology for surface science calculations. It identifies surface atoms based on coordination analysis relative
-    to the bulk structure, tiles the slab to meet minimum lateral size requirements for accurate periodic calculations,
-    and optionally applies constraints to fix bulk-like atoms during optimization. This processing is essential for
-    creating production-ready slab models for catalysis and surface chemistry studies. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use after generating raw slabs that need surface atom identification and proper sizing
-    - Essential for preparing slabs for DFT calculations or molecular dynamics simulations
-    - Required when you need to distinguish between surface and bulk atoms in the slab
-    - Use when raw slabs are too small for accurate periodic boundary calculations
-    - Avoid for slabs that have already been properly processed and tagged
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Loads both slab and bulk structures from CIF strings
-    - Standardizes the bulk structure for consistent coordination reference
-    - Uses Voronoi analysis to identify surface atoms based on coordination differences
-    - Tags atoms as surface (tag=1) or bulk-like (tag=0) based on coordination analysis
-    - Tiles the slab structure to meet minimum lateral size requirements
-    - Applies FixAtoms constraints to bulk-like atoms if requested
-    - Returns the processed slab as a CIF string ready for calculations
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Generate raw slab using create_slab_from_structure_text or enumerate_slabs_text [/PREREQUISITE]
-    2. [PREREQUISITE] Have original bulk structure available for coordination reference [/PREREQUISITE]
-    3. [CURRENT] Apply this tool to process and tag the slab structure [/CURRENT]
-    4. [FOLLOW_UP] Use processed slab for adsorption studies or DFT calculations [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `process_slab_ocdata_style(slab_cif, bulk_cif, 8.0, True)`,
-    `process_slab_ocdata_style(slab_cif, bulk_cif, min_xy_size=10.0)`,
-    `process_slab_ocdata_style(slab_cif, bulk_cif, apply_constraints=False)`,
-    [/SYNTACTICAL]
-
-    Args:
-        slab_cif: [BRIEF] CIF string of the raw slab structure from pymatgen generation. [/BRIEF]
-                 [DETAILED] A properly formatted CIF string containing the raw slab structure that needs
-                 processing. This should be a freshly generated slab from tools like SlabGenerator that
-                 has not yet been tagged or processed for surface calculations. [/DETAILED]
-                 [SYNTACTIC] Format: "Valid CIF format string with slab atomic coordinates" [/SYNTACTIC]
-                 [EXAMPLES] Examples: Output from create_slab_from_structure_text [/EXAMPLES]
-        bulk_cif: [BRIEF] CIF string of the original bulk structure for coordination reference. [/BRIEF]
-                 [DETAILED] The bulk crystal structure that was used to generate the slab. This is required
-                 for accurate surface atom identification as it provides the reference coordination environment
-                 for distinguishing between surface and bulk-like atoms in the slab. [/DETAILED]
-                 [SYNTACTIC] Format: "Valid CIF format string with bulk crystal structure" [/SYNTACTIC]
-                 [EXAMPLES] Examples: Original bulk structure from Materials Project [/EXAMPLES]
-        min_xy_size: [BRIEF] Minimum lateral size in Angstroms for slab tiling. Defaults to 8.0. [/BRIEF]
-                    [DETAILED] The minimum lateral dimension (in XY plane) that the slab should span after
-                    tiling. This ensures adequate periodic boundary conditions for accurate surface calculations.
-                    Larger values provide better convergence but increase computational cost. [/DETAILED]
-                    [SYNTACTIC] Format: positive float representing size in Angstroms [/SYNTACTIC]
-                    [EXAMPLES] Examples: 8.0 (minimal), 10.0 (standard), 12.0 (large) [/EXAMPLES]
-        apply_constraints: [BRIEF] Whether to fix bulk-like atoms with constraints. Defaults to True. [/BRIEF]
-                         [DETAILED] Controls whether to apply FixAtoms constraints to bulk-like atoms (tag=0)
-                         during structural optimization. This is recommended for surface calculations to maintain
-                         the bulk-like character of the slab interior while allowing surface atoms to relax. [/DETAILED]
-                         [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                         [EXAMPLES] Examples: True (recommended), False (all atoms free) [/EXAMPLES]
-
-    Returns:
-        str: [BRIEF] CIF string of the processed slab with surface tagging, tiling, and constraints. [/BRIEF]
-             [DETAILED] A CIF-formatted string containing the fully processed slab structure with surface
-             atoms properly tagged, adequate lateral size for periodic calculations, and optional constraints
-             applied to bulk-like atoms. Ready for surface chemistry calculations. [/DETAILED]
-             [EXAMPLES] Example output: CIF string with tagged atoms and proper dimensions [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When CIF strings are malformed or incompatible structures [/ERROR_WHEN]
-                   [ERROR_DETAILS] Invalid CIF format, mismatched slab-bulk structures, or processing errors [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Verify CIF formats and ensure slab was generated from the provided bulk [/ERROR_RECOVERY]
-        StructureError: [ERROR_WHEN] When surface atom tagging or tiling fails [/ERROR_WHEN]
-                       [ERROR_DETAILS] Coordination analysis failure or insufficient structure size [/ERROR_DETAILS]
-                       [ERROR_RECOVERY] Check structure quality and increase min_xy_size if needed [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Requires the original bulk structure for accurate surface atom identification
-    - May not handle highly defective or reconstructed surfaces properly
-    - Surface tagging accuracy depends on the quality of coordination analysis
-    - Does not account for surface relaxation or reconstruction effects
-    [/LIMITATIONS]
-    """
-    from pymatgen.core import Structure
-    from pymatgen.io.ase import AseAtomsAdaptor
-
-    try:
-        # Load structures
-        slab_struct_pmg = load_structure(slab_cif)
-        slab_atoms_ase = AseAtomsAdaptor.get_atoms(slab_struct_pmg)
-
-        bulk_struct_pmg = load_structure(bulk_cif)
-        # Standardize bulk *before* getting ASE atoms for consistent coordination check
-        standardized_bulk_pmg = standardize_bulk(bulk_struct_pmg)
-        standardized_bulk_ase = AseAtomsAdaptor.get_atoms(standardized_bulk_pmg)
-
-        # 1. Tag Surface Atoms
-        tags = find_surface_atoms_with_voronoi(standardized_bulk_ase, slab_atoms_ase)
-        slab_atoms_ase.set_tags(tags)
-
-        # 2. Tile the Tagged Slab
-        tiled_atoms_ase = tile_atoms(slab_atoms_ase, min_xy_size)
-
-        # 3. Apply Constraints (Optional)
-        final_atoms_ase = tiled_atoms_ase
-        if apply_constraints:
-            final_atoms_ase = set_fixed_atom_constraints(tiled_atoms_ase)
-
-        # 4. Convert back to CIF
-        final_struct_pmg = Structure.from_ase_atoms(final_atoms_ase)
-        return final_struct_pmg.to(fmt="cif")
-
-    except Exception as e:
-        return f"ERROR: Slab processing failed - {e}"
-
-
-@tool
-def get_symmetrically_distinct_miller_indices_from_bulk(
-    bulk_structure_path_or_string: str, from_path: bool = False, max_miller: int = 2
-) -> list:
-    """[BRIEF] Identify symmetrically distinct Miller indices for a bulk crystal structure using crystallographic symmetry analysis. [/BRIEF]
-
-    [DETAILED] This tool analyzes the crystallographic symmetry of a bulk structure to identify Miller indices
-    that correspond to genuinely different surface orientations. By considering the crystal's space group symmetry,
-    it eliminates redundant Miller indices that would produce equivalent surfaces, significantly reducing the
-    computational burden of systematic surface studies. This is essential for comprehensive surface analysis
-    while avoiding unnecessary calculations on symmetrically equivalent surfaces. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use when planning systematic surface studies across multiple orientations
-    - Essential before generating multiple slabs to avoid redundant calculations
-    - Required for high-throughput surface screening studies
-    - Use when you need to understand available surface orientations for a material
-    - Avoid when you only need specific, known Miller indices
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Loads the bulk crystal structure from CIF file or string
-    - Analyzes the crystal's space group symmetry operations
-    - Generates all possible Miller indices up to the specified maximum
-    - Applies symmetry operations to identify equivalent Miller indices
-    - Returns only the symmetrically distinct indices that produce unique surfaces
-    - Uses pymatgen's crystallographic analysis for accurate symmetry determination
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Have bulk crystal structure available as CIF file or string [/PREREQUISITE]
-    2. [CURRENT] Apply this tool to identify distinct Miller indices [/CURRENT]
-    3. [FOLLOW_UP] Use output with enumerate_slabs_for_list_of_miller_index for efficient slab generation [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `get_symmetrically_distinct_miller_indices_from_bulk(cif_string, False, 2)`,
-    `get_symmetrically_distinct_miller_indices_from_bulk("structure.cif", True, 3)`,
-    `get_symmetrically_distinct_miller_indices_from_bulk(cif_string)`,
-    [/SYNTACTICAL]
-
-    Args:
-        bulk_structure_path_or_string: [BRIEF] Path to CIF file or CIF string of the bulk structure. [/BRIEF]
-                                     [DETAILED] Either a file path to a CIF file containing the bulk crystal
-                                     structure, or a CIF-formatted string. The structure should be a well-defined
-                                     bulk crystal with proper space group information for accurate symmetry analysis. [/DETAILED]
-                                     [SYNTACTIC] Format: "file/path/to/structure.cif" or "CIF format string" [/SYNTACTIC]
-                                     [EXAMPLES] Examples: "bulk_structure.cif", CIF string content [/EXAMPLES]
-        from_path: [BRIEF] Boolean indicating if input is a file path. Defaults to False. [/BRIEF]
-                  [DETAILED] Controls whether the first argument should be treated as a file path (True)
-                  or as a CIF string content (False). This allows the tool to handle both file-based
-                  and string-based inputs flexibly. [/DETAILED]
-                  [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                  [EXAMPLES] Examples: True (file path), False (CIF string) [/EXAMPLES]
-        max_miller: [BRIEF] Maximum Miller index to consider. Defaults to 2. [/BRIEF]
-                   [DETAILED] The maximum value for Miller indices (h, k, l) to consider in the analysis.
-                   Higher values include more high-index surfaces but increase computational requirements.
-                   Values of 1-3 are typically sufficient for most surface studies. [/DETAILED]
-                   [SYNTACTIC] Format: positive integer (1, 2, or 3) [/SYNTACTIC]
-                   [EXAMPLES] Examples: 1 (low-index only), 2 (standard), 3 (high-index) [/EXAMPLES]
-
-    Returns:
-        list: [BRIEF] List of symmetrically distinct Miller indices as tuples. [/BRIEF]
-              [DETAILED] A list of tuples, each containing three integers representing Miller indices
-              (h, k, l) that correspond to crystallographically distinct surface orientations. Each
-              tuple represents a unique surface that cannot be obtained from others by symmetry operations. [/DETAILED]
-              [EXAMPLES] Example output: [(1,0,0), (1,1,0), (1,1,1)] [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When CIF structure is malformed or max_miller is invalid [/ERROR_WHEN]
-                   [ERROR_DETAILS] Invalid CIF format, file not found, or max_miller outside valid range [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Verify CIF format and ensure max_miller is 1, 2, or 3 [/ERROR_RECOVERY]
-        StructureError: [ERROR_WHEN] When symmetry analysis fails [/ERROR_WHEN]
-                       [ERROR_DETAILS] Unclear space group or problematic crystal structure [/ERROR_DETAILS]
-                       [ERROR_RECOVERY] Check structure quality and space group assignment [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Accuracy depends on correct space group assignment in the input structure
-    - May not account for surface-specific symmetry breaking
-    - Limited to crystalline materials with well-defined space groups
-    - Does not consider surface stability or feasibility
-    [/LIMITATIONS]
-    """
-
-    from pymatgen.core.surface import get_symmetrically_distinct_miller_indices
-
-    bulk_structure = load_structure(bulk_structure_path_or_string, from_path)
-
-    return get_symmetrically_distinct_miller_indices(bulk_structure, max_miller)
-
-
-@tool
-def enumerate_all_possible_miller_indices(max_miller: int = 2) -> list:
-    """[BRIEF] Generate all possible Miller indices up to a specified maximum value for comprehensive surface enumeration. [/BRIEF]
-
-    [DETAILED] This tool systematically generates all possible combinations of Miller indices (h, k, l) up to a
-    specified maximum value, providing a complete set of surface orientations for systematic surface studies.
-    Unlike symmetry-based filtering, this tool produces all mathematically possible Miller indices regardless
-    of crystallographic equivalence, useful for comprehensive surface exploration or when symmetry information
-    is unavailable. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use when you need a complete enumeration of all possible surface orientations
-    - Helpful for systematic surface studies without symmetry constraints
-    - Use when symmetry information is unavailable or unreliable
-    - Good for initial exploration of surface space before symmetry filtering
-    - Avoid when you only need symmetrically distinct surfaces (use get_symmetrically_distinct_miller_indices_from_bulk)
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Generates all combinations of integers from 0 to max_miller for (h, k, l)
-    - Excludes the invalid (0, 0, 0) combination
-    - Returns a comprehensive list of all possible Miller index combinations
-    - No crystallographic analysis or symmetry filtering applied
-    - Simple mathematical enumeration approach
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [CURRENT] Apply this tool to generate all possible Miller indices [/CURRENT]
-    2. [FOLLOW_UP] Filter results using get_symmetrically_distinct_miller_indices_from_bulk if needed [/FOLLOW_UP]
-    3. [FOLLOW_UP] Use selected indices with enumerate_slabs_for_list_of_miller_index [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `enumerate_all_possible_miller_indices(2)`,
-    `enumerate_all_possible_miller_indices(3)`,
-    `enumerate_all_possible_miller_indices()`,  # Uses default max_miller=2
-    [/SYNTACTICAL]
-
-    Args:
-        max_miller: [BRIEF] Maximum Miller index value to consider. Defaults to 2. [/BRIEF]
-                   [DETAILED] The maximum value for each Miller index component (h, k, l). The tool will
-                   generate all combinations where each component is between 0 and max_miller inclusive.
-                   Higher values produce more surface orientations but increase the total number exponentially. [/DETAILED]
-                   [SYNTACTIC] Format: positive integer (typically 1, 2, or 3) [/SYNTACTIC]
-                   [EXAMPLES] Examples: 1 (27 combinations), 2 (125 combinations), 3 (343 combinations) [/EXAMPLES]
-
-    Returns:
-        list: [BRIEF] List of all possible Miller indices as tuples. [/BRIEF]
-              [DETAILED] A comprehensive list of tuples, each containing three integers representing
-              Miller indices (h, k, l). Includes all mathematical combinations from (0,0,1) to
-              (max_miller, max_miller, max_miller), excluding only the invalid (0,0,0). [/DETAILED]
-              [EXAMPLES] Example output for max_miller=1: [(0,0,1), (0,1,0), (0,1,1), (1,0,0), ...] [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When max_miller is not a positive integer [/ERROR_WHEN]
-                   [ERROR_DETAILS] Negative values or non-integer input for max_miller [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Provide a positive integer value for max_miller [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Does not consider crystallographic symmetry or equivalence
-    - May produce many redundant surface orientations
-    - Number of combinations grows as (max_miller+1)³
-    - No filtering for physically meaningful or stable surfaces
-    [/LIMITATIONS]
-    """
-
-    mill_list = []
-    for i in range(max_miller + 1):
-        for j in range(max_miller + 1):
-            for k in range(max_miller + 1):
-                if i == 0 and j == 0 and k == 0:
-                    continue  # Skip (0,0,0)
-                mill_list.append((i, j, k))
-    return mill_list
-
-
-@tool
-def find_all_unique_slabs_upto_millerindex_to_file(
-    bulk_structure_path_or_string: str,
-    out_put_path: str,
-    from_path: bool = False,
-    max_index: int = 2,
-    min_slab_size: float = 8,
-    min_vacuum_size: float = 15,
-    center_slab: bool = True,
-    max_normal_search: int = 10,
-) -> str:
-    """[BRIEF] Generate all unique slab structures up to specified Miller indices and save comprehensive results to JSON file. [/BRIEF]
-
-    [DETAILED] This tool performs exhaustive slab generation for a bulk structure, creating all possible unique
-    surface slabs up to a specified maximum Miller index. It systematically explores different surface orientations
-    and terminations, providing a comprehensive database of surface structures for high-throughput surface studies.
-    The results include structural information, surface area, thickness, and other properties essential for
-    surface analysis and screening studies. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use for comprehensive surface screening and high-throughput studies
-    - Essential when you need a complete database of surface structures
-    - Use when planning systematic surface property calculations
-    - Recommended for materials with unknown optimal surface orientations
-    - Avoid for targeted studies where specific surfaces are already known
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Loads the bulk structure from file or CIF string
-    - Uses pymatgen's generate_all_slabs for comprehensive slab generation
-    - Systematically explores all Miller indices up to max_index
-    - Generates multiple terminations for each Miller index
-    - Collects detailed information including surface area, thickness, and atom count
-    - Saves all results to a structured JSON file for easy access and analysis
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Have bulk crystal structure available as file or string [/PREREQUISITE]
-    2. [CURRENT] Apply this tool to generate comprehensive slab database [/CURRENT]
-    3. [FOLLOW_UP] Use select_slabs_with_strategy_to_file to filter results [/FOLLOW_UP]
-    4. [FOLLOW_UP] Process selected slabs with surface analysis tools [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `find_all_unique_slabs_upto_millerindex_to_file(cif_string, "slabs.json", False, 2, 8, 15, True, 10)`,
-    `find_all_unique_slabs_upto_millerindex_to_file("bulk.cif", "output.json", True)`,
-    `find_all_unique_slabs_upto_millerindex_to_file(cif_string, "slabs.json")`,
-    [/SYNTACTICAL]
-
-    Args:
-        bulk_structure_path_or_string: [BRIEF] Path to CIF file or CIF string of the bulk structure. [/BRIEF]
-                                     [DETAILED] Either a file path to a CIF file or a CIF-formatted string
-                                     containing the bulk crystal structure. This structure will be systematically
-                                     cleaved along all possible Miller indices to generate surface slabs. [/DETAILED]
-                                     [SYNTACTIC] Format: "file/path/to/structure.cif" or "CIF format string" [/SYNTACTIC]
-                                     [EXAMPLES] Examples: "bulk_structure.cif", CIF string content [/EXAMPLES]
-        out_put_path: [BRIEF] Path where the JSON output file will be saved. [/BRIEF]
-                     [DETAILED] Full file path for saving the comprehensive slab database as a JSON file.
-                     The file will contain structured data with slab IDs, Miller indices, termination information,
-                     CIF structures, and calculated properties for each generated slab. [/DETAILED]
-                     [SYNTACTIC] Format: "path/to/output_file.json" [/SYNTACTIC]
-                     [EXAMPLES] Examples: "slabs_database.json", "output/all_slabs.json" [/EXAMPLES]
-        from_path: [BRIEF] Boolean indicating if input is a file path. Defaults to False. [/BRIEF]
-                  [DETAILED] Controls whether the bulk_structure_path_or_string parameter should be treated
-                  as a file path (True) or as a CIF string content (False). This allows flexible input handling
-                  for both file-based and string-based workflows. [/DETAILED]
-                  [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                  [EXAMPLES] Examples: True (file path), False (CIF string) [/EXAMPLES]
-        max_index: [BRIEF] Maximum Miller index to consider. Defaults to 2. [/BRIEF]
-                  [DETAILED] The maximum value for Miller indices (h, k, l) to include in the systematic
-                  slab generation. Higher values include more surface orientations but significantly increase
-                  computational requirements and the number of generated slabs. [/DETAILED]
-                  [SYNTACTIC] Format: positive integer (1, 2, or 3) [/SYNTACTIC]
-                  [EXAMPLES] Examples: 1 (low-index), 2 (standard), 3 (high-index) [/EXAMPLES]
-        min_slab_size: [BRIEF] Minimum slab thickness in Angstroms. Defaults to 8. [/BRIEF]
-                      [DETAILED] The minimum thickness of generated slabs in the direction perpendicular
-                      to the surface plane. This ensures adequate bulk-like character in the slab center
-                      while maintaining the desired surface properties. [/DETAILED]
-                      [SYNTACTIC] Format: positive float representing thickness in Angstroms [/SYNTACTIC]
-                      [EXAMPLES] Examples: 8.0 (minimal), 12.0 (standard), 15.0 (thick) [/EXAMPLES]
-        min_vacuum_size: [BRIEF] Minimum vacuum size in Angstroms. Defaults to 15. [/BRIEF]
-                        [DETAILED] The minimum vacuum space above the surface to prevent interactions
-                        between periodic images. Larger values ensure better isolation but increase
-                        computational cost. The default of 15 Å is suitable for most surface calculations. [/DETAILED]
-                        [SYNTACTIC] Format: positive float representing vacuum thickness in Angstroms [/SYNTACTIC]
-                        [EXAMPLES] Examples: 10.0 (minimal), 15.0 (standard), 20.0 (large) [/EXAMPLES]
-        center_slab: [BRIEF] Whether to center the slab in the vacuum region. Defaults to True. [/BRIEF]
-                    [DETAILED] Controls whether the slab is centered within the vacuum region or positioned
-                    at one end. Centering is generally preferred for surface calculations as it provides
-                    symmetric vacuum on both sides of the slab. [/DETAILED]
-                    [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                    [EXAMPLES] Examples: True (recommended), False (asymmetric) [/EXAMPLES]
-        max_normal_search: [BRIEF] Maximum number of normals to search for slab generation. Defaults to 10. [/BRIEF]
-                          [DETAILED] The maximum number of different normal vectors to consider when searching
-                          for valid slab orientations. Higher values may find more slab orientations but
-                          increase generation time. [/DETAILED]
-                          [SYNTACTIC] Format: positive integer [/SYNTACTIC]
-                          [EXAMPLES] Examples: 5 (fast), 10 (standard), 20 (thorough) [/EXAMPLES]
-
-    Returns:
-        str: [BRIEF] Confirmation message indicating where the slabs data has been written. [/BRIEF]
-             [DETAILED] A status message confirming successful generation and saving of the slab database.
-             The actual slab data is saved to the specified JSON file, which contains detailed information
-             about each generated slab including Miller indices, terminations, and structural properties. [/DETAILED]
-             [EXAMPLES] Example output: "Slabs data written to slabs_database.json" [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When input parameters are invalid or CIF structure is malformed [/ERROR_WHEN]
-                   [ERROR_DETAILS] Invalid CIF format, negative size parameters, or file path issues [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Verify CIF format and ensure all size parameters are positive [/ERROR_RECOVERY]
-        IOError: [ERROR_WHEN] When output file cannot be written [/ERROR_WHEN]
-                [ERROR_DETAILS] Insufficient permissions or invalid output path [/ERROR_DETAILS]
-                [ERROR_RECOVERY] Check file permissions and ensure output directory exists [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Can generate very large numbers of slabs for high max_index values
-    - Does not filter for surface stability or feasibility
-    - May include slabs with unusual or unstable terminations
-    - Processing time increases significantly with max_index
-    [/LIMITATIONS]
-    """
-
-    from pymatgen.core.surface import generate_all_slabs
-
-    bulk_structure = load_structure(bulk_structure_path_or_string, from_path)
-
-    slabs = generate_all_slabs(
-        bulk_structure,
-        max_index=max_index,
-        min_slab_size=min_slab_size,
-        min_vacuum_size=min_vacuum_size,
-        center_slab=center_slab,
-        max_normal_search=max_normal_search,
-    )
-    slabs_dict = {}
-    for i, slab in enumerate(slabs):
-        slab_id = (
-            f"{slab.miller_index[0]}{slab.miller_index[1]}{slab.miller_index[2]}_{i}"
-        )
-        slabs_dict[slab_id] = {
-            "miller_index": slab.miller_index,
-            "termination": i,
-            "cif": slab.to(fmt="cif"),
-            "area": slab.surface_area,
-            "num_sites": len(slab),
-            "slab_thickness": slab.thickness,
-        }
-
-    with Path(out_put_path).open("w") as f:
-        json.dump(slabs_dict, f, indent=2)
-    return f"Slabs data written to {out_put_path}"
-
-
-@tool
-def enumerate_slabs_for_list_of_miller_index(
-    bulk_structure_path_or_string: str,
-    from_path: bool = False,
-    miller_index_list: list[tuple] | None = None,
-    min_slab_size: float = 12,
-    min_vacuum_size: float = 5,
-) -> str:
-    """[BRIEF] Generate slab structures for a specific list of Miller indices with customizable dimensions. [/BRIEF]
-
-    [DETAILED] This tool generates surface slab structures for a user-specified list of Miller indices,
-    providing precise control over which surface orientations to create. Unlike exhaustive generation methods,
-    this tool allows targeted slab creation for specific crystallographic planes, making it ideal for focused
-    surface studies or when specific surface orientations are known to be of interest. Each slab is properly
-    oriented and sized according to the specified parameters. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use when you need slabs for specific, known Miller indices
-    - Ideal for targeted surface studies rather than comprehensive screening
-    - Use after identifying interesting surfaces from symmetry analysis
-    - Recommended when computational resources are limited
-    - Avoid for exhaustive surface exploration (use find_all_unique_slabs_upto_millerindex_to_file)
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Loads the bulk structure from file or CIF string
-    - Iterates through the provided list of Miller indices
-    - Creates SlabGenerator for each Miller index with specified dimensions
-    - Generates all possible terminations for each Miller index
-    - Sorts and orients each slab with surface normal along c-axis
-    - Returns a JSON string containing all generated slabs with identifiers
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Identify Miller indices using get_symmetrically_distinct_miller_indices_from_bulk [/PREREQUISITE]
-    2. [CURRENT] Apply this tool to generate slabs for selected Miller indices [/CURRENT]
-    3. [FOLLOW_UP] Use select_slabs_with_strategy to filter results if needed [/FOLLOW_UP]
-    4. [FOLLOW_UP] Process slabs with surface analysis or adsorption tools [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `enumerate_slabs_for_list_of_miller_index(cif_string, False, [(1,1,1), (1,0,0)], 12, 5)`,
-    `enumerate_slabs_for_list_of_miller_index("bulk.cif", True, [(1,1,1), (1,1,0)])`,
-    `enumerate_slabs_for_list_of_miller_index(cif_string, miller_index_list=[(2,1,1)])`,
-    [/SYNTACTICAL]
-
-    Args:
-        bulk_structure_path_or_string: [BRIEF] Path to CIF file or CIF string of the bulk structure. [/BRIEF]
-                                     [DETAILED] Either a file path to a CIF file or a CIF-formatted string
-                                     containing the bulk crystal structure that will be cleaved along the
-                                     specified Miller indices to create surface slabs. [/DETAILED]
-                                     [SYNTACTIC] Format: "file/path/to/structure.cif" or "CIF format string" [/SYNTACTIC]
-                                     [EXAMPLES] Examples: "bulk_structure.cif", CIF string content [/EXAMPLES]
-        from_path: [BRIEF] Boolean indicating if input is a file path. Defaults to False. [/BRIEF]
-                  [DETAILED] Controls whether the bulk_structure_path_or_string parameter should be treated
-                  as a file path (True) or as a CIF string content (False). This allows flexible input handling
-                  for both file-based and string-based workflows. [/DETAILED]
-                  [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                  [EXAMPLES] Examples: True (file path), False (CIF string) [/EXAMPLES]
-        miller_index_list: [BRIEF] List of Miller indices to generate slabs for. Required parameter. [/BRIEF]
-                          [DETAILED] A list of tuples, each containing three integers representing Miller
-                          indices (h, k, l) for the desired surface orientations. Each tuple specifies a
-                          crystallographic plane along which the bulk structure will be cleaved. [/DETAILED]
-                          [SYNTACTIC] Format: list of tuples [(h1,k1,l1), (h2,k2,l2), ...] [/SYNTACTIC]
-                          [EXAMPLES] Examples: [(1,1,1), (1,0,0), (1,1,0)], [(2,1,1), (1,2,1)] [/EXAMPLES]
-        min_slab_size: [BRIEF] Minimum slab thickness in Angstroms. Defaults to 12. [/BRIEF]
-                      [DETAILED] The minimum thickness of generated slabs in the direction perpendicular
-                      to the surface plane. This ensures adequate bulk-like character in the slab center
-                      while maintaining the desired surface properties. [/DETAILED]
-                      [SYNTACTIC] Format: positive float representing thickness in Angstroms [/SYNTACTIC]
-                      [EXAMPLES] Examples: 8.0 (minimal), 12.0 (standard), 15.0 (thick) [/EXAMPLES]
-        min_vacuum_size: [BRIEF] Minimum vacuum size in Angstroms. Defaults to 5. [/BRIEF]
-                        [DETAILED] The minimum vacuum space above the surface to prevent interactions
-                        between periodic images in surface calculations. Smaller values are acceptable
-                        for initial slab generation, but larger values may be needed for accurate calculations. [/DETAILED]
-                        [SYNTACTIC] Format: positive float representing vacuum thickness in Angstroms [/SYNTACTIC]
-                        [EXAMPLES] Examples: 5.0 (minimal), 10.0 (standard), 15.0 (large) [/EXAMPLES]
-
-    Returns:
-        str: [BRIEF] JSON string containing all generated slabs with identifiers and CIF structures. [/BRIEF]
-             [DETAILED] A JSON-formatted string containing a dictionary where keys are slab identifiers
-             (e.g., "slab_0_(1,1,1)") and values are CIF strings of the corresponding slab structures.
-             Each slab is properly oriented and sized according to the specified parameters. [/DETAILED]
-             [EXAMPLES] Example output: {"slab_0_(1,1,1)": "CIF string", "slab_1_(1,0,0)": "CIF string"} [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When miller_index_list is None or contains invalid indices [/ERROR_WHEN]
-                   [ERROR_DETAILS] Missing Miller indices list or invalid tuple format [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Provide a valid list of Miller index tuples [/ERROR_RECOVERY]
-        StructureError: [ERROR_WHEN] When CIF structure is malformed or slab generation fails [/ERROR_WHEN]
-                       [ERROR_DETAILS] Invalid CIF format or incompatible Miller indices [/ERROR_DETAILS]
-                       [ERROR_RECOVERY] Verify CIF format and ensure Miller indices are valid for the structure [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Requires explicit specification of Miller indices
-    - Does not perform symmetry analysis or filtering
-    - May generate multiple terminations per Miller index
-    - Does not optimize slab dimensions for specific surfaces
-    [/LIMITATIONS]
-    """
-
-    import json
-
-    from pymatgen.core import Structure
-    from pymatgen.core.surface import SlabGenerator
-
-    if miller_index_list is None:
-        raise ValueError("Miller indexes should be defined")
-
-    if from_path:
-        bulk_structure = Structure.from_file(bulk_structure_path_or_string)
-    else:
-        bulk_structure = Structure.from_str(bulk_structure_path_or_string, fmt="cif")
-
-    slabs_dict = {}
-    for millers in miller_index_list:
-        slab_gen = SlabGenerator(
-            bulk_structure, millers, min_slab_size, min_vacuum_size
-        )
-        slabs = slab_gen.get_slabs()  # returns a list of Slab objects
-        for i, slab in enumerate(slabs):
-            # We use get_orthogonal_c_slab() ensures that the slab lattice is reoriented in c axis for easier adsorption placement.
-            # get_sorted_structure() variations in atom ordering that might occur due to how the slab was originally created.
-            slab_clean = (
-                slab.get_sorted_structure()
-                # slab.get_orthogonal_c_slab().get_sorted_structure()
-            )
-            slabs_dict[f"slab_{i}_{millers}"] = slab_clean.to(fmt="cif")
-
-    return json.dumps(slabs_dict, indent=2)
-
-
-@tool
-def select_slabs_with_strategy(
-    slabs_json: str,
-    selection_strategy: str = "diverse_miller",
-    max_slabs_per_polymorph: int = 3,
-) -> str:
-    """[BRIEF] Select a subset of slabs from a collection using intelligent filtering strategies for focused studies. [/BRIEF]
-
-    [DETAILED] This tool applies intelligent selection strategies to filter large collections of slab structures,
-    choosing the most relevant slabs based on specified criteria. It supports multiple selection strategies including
-    diversity-based selection (diverse Miller indices), property-based selection (surface area, coordination),
-    and custom filtering approaches. This is essential for managing large slab databases and focusing computational
-    resources on the most promising surface structures. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use after generating large collections of slabs to focus on most relevant structures
-    - Essential for managing computational resources in high-throughput studies
-    - Use when you need representative samples from comprehensive slab databases
-    - Recommended for initial screening before expensive calculations
-    - Avoid when you need all generated slabs for comprehensive analysis
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Parses the input JSON containing slab data and properties
-    - Applies the selected strategy to rank and filter slabs
-    - diverse_miller: Selects slabs with different Miller indices for surface diversity
-    - large_surface: Prioritizes slabs with the largest surface areas
-    - high_coordination: Selects slabs with higher numbers of surface sites
-    - Returns a filtered JSON with the selected slabs and their properties
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Generate slab collection using find_all_unique_slabs_upto_millerindex_to_file [/PREREQUISITE]
-    2. [CURRENT] Apply this tool to select most relevant slabs [/CURRENT]
-    3. [FOLLOW_UP] Use selected slabs for detailed surface analysis or adsorption studies [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `select_slabs_with_strategy(slabs_json, "diverse_miller", 3)`,
-    `select_slabs_with_strategy(slabs_json, "large_surface", 5)`,
-    `select_slabs_with_strategy(slabs_json, "high_coordination")`,
-    [/SYNTACTICAL]
-
-    Args:
-        slabs_json: [BRIEF] JSON string containing slab data with properties and structures. [/BRIEF]
-                   [DETAILED] A JSON-formatted string containing a dictionary of slab data, where each entry
-                   includes slab properties such as Miller indices, surface area, number of sites, and CIF
-                   structures. This should be the output from slab generation tools. [/DETAILED]
-                   [SYNTACTIC] Format: "JSON string with slab data dictionary" [/SYNTACTIC]
-                   [EXAMPLES] Examples: Output from find_all_unique_slabs_upto_millerindex_to_file [/EXAMPLES]
-        selection_strategy: [BRIEF] Strategy for slab selection. Defaults to "diverse_miller". [/BRIEF]
-                          [DETAILED] The selection strategy to apply for filtering slabs. Options include:
-                          "diverse_miller" for different Miller indices, "large_surface" for largest surface areas,
-                          "high_coordination" for highest number of surface sites. Each strategy focuses on
-                          different aspects of surface structure importance. [/DETAILED]
-                          [SYNTACTIC] Format: string from ["diverse_miller", "large_surface", "high_coordination"] [/SYNTACTIC]
-                          [EXAMPLES] Examples: "diverse_miller", "large_surface", "high_coordination" [/EXAMPLES]
-        max_slabs_per_polymorph: [BRIEF] Maximum number of slabs to select. Defaults to 3. [/BRIEF]
-                               [DETAILED] The maximum number of slabs to include in the filtered selection.
-                               This controls the size of the final slab collection and helps manage computational
-                               resources for subsequent analysis. [/DETAILED]
-                               [SYNTACTIC] Format: positive integer [/SYNTACTIC]
-                               [EXAMPLES] Examples: 3 (focused), 5 (balanced), 10 (comprehensive) [/EXAMPLES]
-
-    Returns:
-        str: [BRIEF] JSON string containing the selected slabs with their properties and structures. [/BRIEF]
-             [DETAILED] A JSON-formatted string containing the filtered slab collection, with the same structure
-             as the input but containing only the selected slabs according to the specified strategy and
-             maximum count. Each selected slab retains all its original properties and CIF structure. [/DETAILED]
-             [EXAMPLES] Example output: Filtered JSON with selected slabs and their complete data [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When slabs_json is malformed or selection_strategy is invalid [/ERROR_WHEN]
-                   [ERROR_DETAILS] Invalid JSON format or unsupported selection strategy [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Verify JSON format and use valid selection strategy [/ERROR_RECOVERY]
-        KeyError: [ERROR_WHEN] When required slab properties are missing from the input data [/ERROR_WHEN]
-                 [ERROR_DETAILS] Missing miller_index, area, or num_sites properties [/ERROR_DETAILS]
-                 [ERROR_RECOVERY] Ensure input JSON contains all required slab properties [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Selection strategies are predefined and may not cover all use cases
-    - Does not consider surface stability or chemical properties
-    - May not account for computational cost differences between slabs
-    - Selection is based on structural properties rather than target applications
-    [/LIMITATIONS]
-    """
-    slabs = json.loads(slabs_json)
-
-    if selection_strategy == "diverse_miller":
-        # Select slabs with different Miller indices
-        selected = {}
-        miller_indices_seen = set()
-        for slab_id, slab_data in slabs.items():
-            miller_tuple = tuple(slab_data["miller_index"])
-            if (
-                miller_tuple not in miller_indices_seen
-                and len(selected) < max_slabs_per_polymorph
-            ):
-                selected[slab_id] = slab_data
-                miller_indices_seen.add(miller_tuple)
-
-    elif selection_strategy == "large_surface":
-        # Select slabs with largest surface areas
-        sorted_slabs = sorted(slabs.items(), key=lambda x: x[1]["area"], reverse=True)
-        selected = dict(sorted_slabs[:max_slabs_per_polymorph])
-
-    elif selection_strategy == "high_coordination":
-        # Select slabs with higher number of surface sites (proxy for coordination)
-        sorted_slabs = sorted(
-            slabs.items(), key=lambda x: x[1]["num_sites"], reverse=True
-        )
-        selected = dict(sorted_slabs[:max_slabs_per_polymorph])
-
-    else:
-        # Default: take first max_slabs_per_polymorph
-        selected = dict(list(slabs.items())[:max_slabs_per_polymorph])
-
-    return json.dumps(selected, indent=2)
-
-
-@tool
-def select_slabs_with_strategy_to_file(
-    slabs_data: str,
-    save_path: str,
-    selection_strategy: str = "diverse_miller",
-    max_slabs_per_polymorph: int = 3,
-    is_path: bool = False,
-) -> str:
-    """[BRIEF] Select slabs using filtering strategies and save results to file for persistent storage and workflow integration. [/BRIEF]
-
-    [DETAILED] This tool extends the slab selection functionality by providing file-based input/output capabilities,
-    allowing for persistent storage of selection results and better integration with file-based computational workflows.
-    It applies the same intelligent selection strategies as the base selection tool but handles file operations
-    automatically, making it ideal for batch processing and pipeline integration. [/DETAILED]
-
-    [PROCEDURAL] When to use this tool:
-    - Use when you need to save selection results for later use or sharing
-    - Essential for batch processing and automated workflows
-    - Use when working with large slab databases stored in files
-    - Recommended for pipeline integration and reproducible research
-    - Avoid for one-time selections that don't need persistent storage
-    [/PROCEDURAL]
-
-    [CONTEXTUAL] How this tool works:
-    - Loads slab data from JSON string or file path based on is_path parameter
-    - Applies the specified selection strategy to filter slabs
-    - Saves the filtered results to a new JSON file at the specified path
-    - Maintains all original slab properties and structures in the output
-    - Returns the path to the saved file for further workflow integration
-    [/CONTEXTUAL]
-
-    [WORKFLOW_INTEGRATION] Typical workflow integration example:
-    1. [PREREQUISITE] Generate slab database file using find_all_unique_slabs_upto_millerindex_to_file [/PREREQUISITE]
-    2. [CURRENT] Apply this tool to select and save filtered slabs [/CURRENT]
-    3. [FOLLOW_UP] Use saved selection file for subsequent surface analysis workflows [/FOLLOW_UP]
-    [/WORKFLOW_INTEGRATION]
-
-    [SYNTACTICAL] Usage examples:
-    `select_slabs_with_strategy_to_file("slabs.json", "selected.json", "diverse_miller", 3, True)`,
-    `select_slabs_with_strategy_to_file(slabs_json_string, "output.json", "large_surface", 5, False)`,
-    `select_slabs_with_strategy_to_file("input.json", "filtered.json", is_path=True)`,
-    [/SYNTACTICAL]
-
-    Args:
-        slabs_data: [BRIEF] JSON string or file path containing slab data. [/BRIEF]
-                   [DETAILED] Either a JSON-formatted string containing slab data or a file path to a JSON
-                   file with slab information. The data should include slab properties such as Miller indices,
-                   surface area, number of sites, and CIF structures. [/DETAILED]
-                   [SYNTACTIC] Format: "JSON string" or "path/to/slabs.json" [/SYNTACTIC]
-                   [EXAMPLES] Examples: JSON string content, "slabs_database.json" [/EXAMPLES]
-        save_path: [BRIEF] Path where the selected slabs JSON file will be saved. [/BRIEF]
-                  [DETAILED] Full file path for saving the filtered slab collection as a JSON file.
-                  The file will contain the selected slabs with all their original properties and
-                  structures, ready for further analysis or processing. [/DETAILED]
-                  [SYNTACTIC] Format: "path/to/output_file.json" [/SYNTACTIC]
-                  [EXAMPLES] Examples: "selected_slabs.json", "output/filtered_slabs.json" [/EXAMPLES]
-        selection_strategy: [BRIEF] Strategy for slab selection. Defaults to "diverse_miller". [/BRIEF]
-                          [DETAILED] The selection strategy to apply for filtering slabs. Options include:
-                          "diverse_miller" for different Miller indices, "large_surface" for largest surface areas,
-                          "high_coordination" for highest number of surface sites. [/DETAILED]
-                          [SYNTACTIC] Format: string from ["diverse_miller", "large_surface", "high_coordination"] [/SYNTACTIC]
-                          [EXAMPLES] Examples: "diverse_miller", "large_surface", "high_coordination" [/EXAMPLES]
-        max_slabs_per_polymorph: [BRIEF] Maximum number of slabs to select. Defaults to 3. [/BRIEF]
-                               [DETAILED] The maximum number of slabs to include in the filtered selection.
-                               This controls the size of the final slab collection and helps manage computational
-                               resources for subsequent analysis. [/DETAILED]
-                               [SYNTACTIC] Format: positive integer [/SYNTACTIC]
-                               [EXAMPLES] Examples: 3 (focused), 5 (balanced), 10 (comprehensive) [/EXAMPLES]
-        is_path: [BRIEF] Whether slabs_data is a file path. Defaults to False. [/BRIEF]
-                [DETAILED] Controls whether the slabs_data parameter should be treated as a file path (True)
-                or as a JSON string content (False). This allows flexible input handling for both
-                file-based and string-based workflows. [/DETAILED]
-                [SYNTACTIC] Format: boolean value (True/False) [/SYNTACTIC]
-                [EXAMPLES] Examples: True (file path), False (JSON string) [/EXAMPLES]
-
-    Returns:
-        str: [BRIEF] Path to the saved JSON file containing selected slabs. [/BRIEF]
-             [DETAILED] The file path where the filtered slab collection has been saved. This path can be
-             used for subsequent workflow steps or for loading the selected slabs in other tools. [/DETAILED]
-             [EXAMPLES] Example output: "path/to/selected_slabs.json" [/EXAMPLES]
-
-    [RAISES] Exceptions:
-        ValueError: [ERROR_WHEN] When input data is malformed or selection strategy is invalid [/ERROR_WHEN]
-                   [ERROR_DETAILS] Invalid JSON format, unsupported selection strategy, or file reading errors [/ERROR_DETAILS]
-                   [ERROR_RECOVERY] Verify JSON format and use valid selection strategy [/ERROR_RECOVERY]
-        IOError: [ERROR_WHEN] When file operations fail [/ERROR_WHEN]
-                [ERROR_DETAILS] Cannot read input file or write to output path [/ERROR_DETAILS]
-                [ERROR_RECOVERY] Check file permissions and ensure paths are valid [/ERROR_RECOVERY]
-    [/RAISES]
-
-    [LIMITATIONS] Known limitations:
-    - Same selection strategy limitations as the base selection tool
-    - File operations may fail with insufficient permissions
-    - Large files may consume significant memory during processing
-    - No validation of file formats before processing
-    [/LIMITATIONS]
-    """
-    # Load slabs data
-    if is_path:
-        with Path(slabs_data).open("r") as f:
-            slabs = json.loads(f.read())
-    else:
-        slabs = json.loads(slabs_data)
-
-    if selection_strategy == "diverse_miller":
-        # Select slabs with different Miller indices
-        selected = {}
-        miller_indices_seen = set()
-        for slab_id, slab_data in slabs.items():
-            miller_tuple = tuple(slab_data["miller_index"])
-            if (
-                miller_tuple not in miller_indices_seen
-                and len(selected) < max_slabs_per_polymorph
-            ):
-                selected[slab_id] = slab_data
-                miller_indices_seen.add(miller_tuple)
-
-    elif selection_strategy == "large_surface":
-        # Select slabs with largest surface areas
-        sorted_slabs = sorted(slabs.items(), key=lambda x: x[1]["area"], reverse=True)
-        selected = dict(sorted_slabs[:max_slabs_per_polymorph])
-
-    elif selection_strategy == "high_coordination":
-        # Select slabs with higher number of surface sites (proxy for coordination)
-        sorted_slabs = sorted(
-            slabs.items(), key=lambda x: x[1]["num_sites"], reverse=True
-        )
-        selected = dict(sorted_slabs[:max_slabs_per_polymorph])
-
-    else:
-        # Default: take first max_slabs_per_polymorph
-        selected = dict(list(slabs.items())[:max_slabs_per_polymorph])
-
-    # Save to file
-    with Path(save_path).open("w") as f:
-        json.dump(selected, f, indent=2)
-
-    return save_path
-
-
 @tool
 def generate_adsorbate_slab_configs(
     slab_cif: str, adsorbate_cif: str, adsorption_sites_json: str, height: float = 1.8
@@ -3305,15 +2430,10 @@ def create_tools() -> dict[str, Tool]:
         "choose_adsorption_site_text": choose_adsorption_site_text,
         "add_adsorbate_to_slab_text": add_adsorbate_to_slab_text,
         "generate_reconstructed_slab": generate_reconstructed_slab,
-        # OCP hard
+        # OCP hard -  Distractors
         "get_bulk_polymorphs_data": get_bulk_polymorphs_data,
         "sort_and_get_first_from_json": sort_and_get_first_from_json,
-        "process_slab_ocdata_style": process_slab_ocdata_style,
-        "get_symmetrically_distinct_miller_indices_from_bulk": get_symmetrically_distinct_miller_indices_from_bulk,
-        "enumerate_all_possible_miller_indices": enumerate_all_possible_miller_indices,
         "find_all_unique_slabs_upto_millerindex": find_all_unique_slabs_upto_millerindex,
-        "find_all_unique_slabs_upto_millerindex_to_file": find_all_unique_slabs_upto_millerindex_to_file,
-        "enumerate_slabs_for_list_of_miller_index": enumerate_slabs_for_list_of_miller_index,
         "generate_adsorbate_slab_configs": generate_adsorbate_slab_configs,
         "get_mp_surface_properties": get_mp_surface_properties,
         "get_mp_thermo_data": get_mp_thermo_data,
